+++ packages/bara/src/index.ts
@@ -1,14 +1,2 @@
-import * as bara from '@bara/core';
-
-const {
-  register,
-  useStream,
-  useTrigger,
-  useEvent,
-  useCondition,
-  useAction,
-} = bara;
-
-export {register, useStream, useTrigger, useEvent, useCondition, useAction};
-
-export default bara;
+export * from '@bara/core'
+export * from '@bara/basics'