+++ packages/basics/src/features/init/event.ts
@@ -1,0 +1,5 @@
+import { createEventType } from '@bara/core'
+
+export interface BaraInit {}
+
+export const ON_INITIALIZED = createEventType('ON_INITIALIZED')