+++ packages/basics/src/features/init/hook.ts
@@ -1,0 +1,7 @@
+import { useInitStream } from './stream'
+import { useInitTrigger } from './trigger'
+
+export function useInit(callback: () => void) {
+  useInitStream()
+  const trigger = useInitTrigger(callback)
+}