+++ packages/core/src/bara.ts
@@ -1,0 +1,130 @@
+import xs from 'xstream'
+
+import { getBaraName } from './helpers/string'
+import {
+  BaraAction,
+  BaraCallbackActionConfig,
+  BaraNormalActionConfig,
+} from './model/action'
+import { AppStream } from './model/app'
+import { BaraConditionConfig } from './model/condition'
+import { BaraEvent, EventType } from './model/event'
+import { BaraStreamConfig, BaraStreamSetup } from './model/stream'
+import {
+  BaraTriggerConfig,
+  BaraTriggerSetup,
+  TriggerEntityType,
+} from './model/trigger'
+
+import { useCallbackActionHook, useNormalActionHook } from './hooks/use-action'
+import { useConditionHook } from './hooks/use-condition'
+import { useEventHook } from './hooks/use-event'
+import { useStreamHook } from './hooks/use-stream'
+import { useTriggerHook } from './hooks/use-trigger'
+
+const bara = (() => {
+  // Main App Stream, get merged with other stream at register step.
+  let appStream: AppStream<any> = xs.never()
+
+  // Handle hook Stream registry
+  const streamRegistry: any[] = []
+  let streamRegistryIndex = 0
+
+  // Handle hook Trigger registry
+  const triggerRegistry: any[] = []
+  let triggerRegistryIndex = 0
+  const triggerConfig: Array<BaraTriggerConfig<any>> = []
+  let triggerConfigIndex = 0
+
+  return {
+    register(app: () => void) {
+      app()
+      return { streamRegistry, triggerRegistry }
+    },
+    useStream<T>(streamSetup: BaraStreamSetup<T>) {
+      streamRegistry[streamRegistryIndex] =
+        streamRegistry[streamRegistryIndex] ||
+        (useStreamHook(streamSetup, streamRegistryIndex) as any)
+
+      // Merge new stream to the main app stream to make global stream
+      appStream = xs.merge(appStream, streamRegistry[streamRegistryIndex]
+        ._$ as AppStream<any>)
+
+      streamRegistryIndex = streamRegistryIndex + 1
+
+      return streamRegistry[streamRegistryIndex - 1]
+    },
+    useTrigger<T>(setup: BaraTriggerSetup<T>) {
+      const config: BaraTriggerConfig<T> = {
+        name: getBaraName(triggerConfigIndex),
+        event: null,
+      }
+
+      // Setup trigger config
+      triggerConfig[triggerConfigIndex] =
+        triggerConfig[triggerConfigIndex] || config
+      const {
+        event: eventSetup,
+        condition: conditionSetup,
+        action: actionSetup,
+      } = setup() // Execute user defined trigger setup function which will assign to the current triggerConfig
+
+      // Done the setup by skip to the next trigger
+      triggerConfigIndex += 1
+
+      // Setup real trigger
+      const currentTrigger =
+        triggerRegistry[triggerRegistryIndex] ||
+        (useTriggerHook(config, triggerRegistryIndex) as any)
+
+      // Attach BaraEvent, BaraCondition, BaraAction with current BaraTrigger
+      const event = currentTrigger.attach(TriggerEntityType.EVENT, eventSetup, [
+        appStream,
+      ])
+      const condition = currentTrigger.attach(
+        TriggerEntityType.CONDITION,
+        conditionSetup,
+        [],
+      )
+      const action = currentTrigger.attach(
+        TriggerEntityType.ACTION,
+        actionSetup,
+        [event, condition],
+      ) // TODO replace event with condition or add condition too
+
+      // Done trigger registering step and skip to next useTrigger function
+      triggerRegistryIndex = triggerRegistryIndex + 1
+
+      return currentTrigger
+    },
+    useEvent<T>(eventType: EventType) {
+      const currentTriggerConfig = triggerConfig[triggerConfigIndex]
+      if (!currentTriggerConfig) {
+        throw new Error(
+          `No trigger is registering at this time. 'useEvent' can only being used in a Bara Trigger.`,
+        )
+      }
+      const eventSetup = useEventHook<T>(currentTriggerConfig, eventType)
+      return eventSetup
+    },
+    useCondition<T>(config: BaraConditionConfig<T>) {
+      const conditionSetup = useConditionHook<T>(config)
+      return conditionSetup
+    },
+    useAction<T>(callback: BaraNormalActionConfig<T>) {
+      const actionSetup = useNormalActionHook(callback)
+      return actionSetup
+    },
+  }
+})()
+
+const {
+  register,
+  useStream,
+  useTrigger,
+  useEvent,
+  useAction,
+  useCondition,
+} = bara
+
+export { register, useStream, useTrigger, useEvent, useAction, useCondition }