+++ packages/core/src/index.ts
@@ -1,1 +1,3 @@
-export * from './core';
+export * from './bara'
+export * from './event'
+export * from './model'