+++ examples/tiktok-app/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/barajs/bara/compare/v2.0.0...v2.1.0) (2019-04-06)
+
+**Note:** Version bump only for package tiktok-app
+
+
+
+
+
+# [2.0.0](https://github.com/barajs/bara/compare/v1.2.0...v2.0.0) (2019-04-04)
+
+**Note:** Version bump only for package tiktok-app
+
+
+
+
+
 # [1.2.0](https://github.com/barajs/bara/compare/v1.1.10...v1.2.0) (2019-03-05)
 
 **Note:** Version bump only for package node-app