+++ packages/bara/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [2.1.0](https://github.com/barajs/bara/compare/v2.0.0...v2.1.0) (2019-04-06)
+
+
+### Features
+
+* **bara:** add @bara/basics to bara package ([88e0f40](https://github.com/barajs/bara/commit/88e0f40))
+* **bara:** export all operators from @bara/core ([c66facd](https://github.com/barajs/bara/commit/c66facd))
+
+
+
+
+
+# [2.0.0](https://github.com/barajs/bara/compare/v1.2.0...v2.0.0) (2019-04-04)
+
+
+### Features
+
+* **bara:** export @bara/core functions ([4dcff4a](https://github.com/barajs/bara/commit/4dcff4a))
+* **bara:** export createEventType function ([36cdea2](https://github.com/barajs/bara/commit/36cdea2))
+
+
+
+
+
 # [1.2.0](https://github.com/barajs/bara/compare/v1.1.10...v1.2.0) (2019-03-05)
 
 **Note:** Version bump only for package bara